Fix: Cambiar parámetros de cantidad de int a float en crearMovimientoConTipo
El método crearMovimientoConTipo recibía cantidades como int, lo que causaba
que valores decimales como 0.0005 se truncaran a 0. Cambié los tipos a float
para soportar correctamente conversiones con decimales.

- Cambiar parámetro $cantidad de int a float
- Cambiar parámetro $stockAnterior de int a float
- Cambiar parámetro $stockNuevo de int a float

+++ app/Http/Controllers/ActualizarStockMasivoController.php
@@ -736,10 +736,10 @@ class ActualizarStockMasivoController extends Controller
     private function crearMovimientoConTipo(
         int $productoId,
         int $stockProductoId,
-        int $cantidad,
+        float $cantidad,
         string $tipo,
-        int $stockAnterior,
-        int $stockNuevo,
+        float $stockAnterior,
+        float $stockNuevo,
         Producto $producto,
         ?string $lote = null
     )
